Refactor CMake
* Simplify CMakeLists
* Move headers file to include directory
