+++ include/search_font.hpp
@@ -1,7 +1,7 @@
 #ifndef KULKI_SEARCH_FONT_HPP
 #define KULKI_SEARCH_FONT_HPP
 
-#include "BabelLib/babel.hpp"
+#include "babel.hpp"
 
 // Finding fonts in disc
 [[nodiscard]] babel::OPT::optional<std::string> search_font() noexcept;
